refactor: remove package repository & move classes to domain
