+++ src/main/java/com/avalon/avalonchat/DataInitializer.java
@@ -9,13 +9,13 @@ import org.springframework.stereotype.Component;
 
 import com.avalon.avalonchat.core.chat.domain.ChatMessageRepository;
 import com.avalon.avalonchat.core.friend.domain.Friend;
-import com.avalon.avalonchat.core.friend.repository.FriendRepository;
+import com.avalon.avalonchat.core.friend.domain.FriendRepository;
 import com.avalon.avalonchat.core.profile.domain.Profile;
-import com.avalon.avalonchat.core.profile.repository.ProfileRepository;
+import com.avalon.avalonchat.core.profile.domain.ProfileRepository;
 import com.avalon.avalonchat.core.user.domain.Email;
 import com.avalon.avalonchat.core.user.domain.Password;
 import com.avalon.avalonchat.core.user.domain.User;
-import com.avalon.avalonchat.core.user.repository.UserRepository;
+import com.avalon.avalonchat.core.user.domain.UserRepository;
 
 import lombok.RequiredArgsConstructor;
