+++ src/main/java/com/avalon/avalonchat/core/friend/domain/FriendRepository.java
@@ -1,4 +1,4 @@
-package com.avalon.avalonchat.core.friend.repository;
+package com.avalon.avalonchat.core.friend.domain;
 
 import java.util.List;
 import java.util.Optional;
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.avalon.avalonchat.core.friend.domain.Friend;
 import com.avalon.avalonchat.core.friend.dto.FriendAddResponse;
 import com.avalon.avalonchat.core.profile.domain.Profile;
