+++ src/main/java/com/avalon/avalonchat/core/profile/domain/ProfileRepository.java
@@ -1,4 +1,4 @@
-package com.avalon.avalonchat.core.profile.repository;
+package com.avalon.avalonchat.core.profile.domain;
 
 import java.util.List;
 import java.util.Optional;
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.avalon.avalonchat.core.profile.domain.Profile;
 import com.avalon.avalonchat.core.profile.dto.ProfileListGetResponse;
 import com.avalon.avalonchat.core.user.domain.User;
