+++ src/main/java/com/avalon/avalonchat/core/user/domain/UserRepository.java
@@ -1,11 +1,8 @@
-package com.avalon.avalonchat.core.user.repository;
+package com.avalon.avalonchat.core.user.domain;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.avalon.avalonchat.core.user.domain.Email;
-import com.avalon.avalonchat.core.user.domain.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 	Optional<User> findByEmail(Email email);
