+++ src/test/java/com/avalon/avalonchat/core/login/application/LoginServiceImplTest.java
@@ -14,11 +14,11 @@ import com.avalon.avalonchat.core.login.dto.EmailFindResponse;
 import com.avalon.avalonchat.core.login.dto.LoginRequest;
 import com.avalon.avalonchat.core.login.dto.LoginResponse;
 import com.avalon.avalonchat.core.profile.domain.Profile;
-import com.avalon.avalonchat.core.profile.repository.ProfileRepository;
-import com.avalon.avalonchat.core.user.domain.User;
-import com.avalon.avalonchat.core.user.dto.SignUpRequest;
-import com.avalon.avalonchat.core.user.repository.UserRepository;
+import com.avalon.avalonchat.core.profile.domain.ProfileRepository;
 import com.avalon.avalonchat.core.user.application.UserServiceImpl;
+import com.avalon.avalonchat.core.user.domain.User;
+import com.avalon.avalonchat.core.user.domain.UserRepository;
+import com.avalon.avalonchat.core.user.dto.SignUpRequest;
 import com.avalon.avalonchat.global.error.exception.BadRequestException;
 import com.avalon.avalonchat.testsupport.DtoFixture;
 import com.avalon.avalonchat.testsupport.Fixture;
